+++ src/quadraticFunction.cpp
@@ -1,5 +1,7 @@
 #include "QuadraticFunction.h"
 
+#include "Stream.h"
+
 QuadraticFunction::QuadraticFunction(float a, float b, float c)
 : a {a}, b {b}, c {c} {}
 
@@ -8,7 +10,7 @@ uint8_t QuadraticFunction::getSlope(float x) const {
   return 2*a*x + b;
 }
 
-float QuadraticFunction::getY(float x) const {
+int16_t QuadraticFunction::getY(int16_t x) const {
   //set x into the function to get y
   return a*(x - b)*(x - b) + c;
 }
